+++ src/components/site/authForm.tsx
@@ -4,9 +4,19 @@ import React from "react";
 import { useTranslations } from "next-intl";
 import { useRouter } from "next-intl/client";
 import { signIn } from "next-auth/react";
-import { Button, Input, Divider } from "@ltopx/lx-ui";
 import toast from "react-hot-toast";
+import { Loader2 } from "lucide-react";
 import { cn } from "@/lib";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import { Button } from "@/components/ui/button";
+import { Label } from "@/components/ui/label";
+import { Input } from "@/components/ui/input";
 import Icon from "@/components/icon";
 import { checkGithub, checkGoogle, checkEmail } from "@/lib/checkEnv";
 
@@ -18,9 +28,9 @@ export default function AuthForm() {
   const inputRef = React.useRef<any>(null);
 
   const [email, setEmail] = React.useState<string>("");
-  const [loadingEmial, setLoadingEmail] = React.useState(false);
   const [loadingGithub, setLoadingGithub] = React.useState(false);
   const [loadingGoogle, setLoadingGoogle] = React.useState(false);
+  const [loadingEmail, setLoadingEmail] = React.useState(false);
 
   const onGithubLogin = async () => {
     setLoadingGithub(true);
@@ -33,7 +43,7 @@ export default function AuthForm() {
   };
 
   const onLogin = async () => {
-    if (loadingEmial) return;
+    if (loadingEmail) return;
 
     if (!email?.trim()) {
       toast.error(tGlobal("please-enter"), { id: "email" });
@@ -68,81 +78,82 @@ export default function AuthForm() {
   };
 
   return (
-    <div
-      className={cn(
-        "shadow-md w-96 z-10 m-auto rounded-xl px-8 py-4 max-w-[calc(100vw-2rem)]",
-        "border border-neutral-200 dark:border-neutral-700/80",
-        "dark:bg-[rgba(24,24,27,.8)]",
-        {
-          "pointer-events-none": loadingEmial || loadingGithub || loadingGoogle,
-        }
-      )}
+    <Card
+      className={cn("w-96 max-w-[calc(100vw-2rem)] relative z-20", {
+        "pointer-events-none": loadingEmail || loadingGithub || loadingGoogle,
+      })}
     >
-      <div className="font-bold mb-2 text-2xl">{tAuth("title")}</div>
-      <div className="text-neutral-500 text-[15px]">
-        {tAuth("continue-gpt")}
-      </div>
-      <div className="flex flex-col mt-8 gap-2">
+      <CardHeader>
+        <CardTitle>{tGlobal("sign-in")}</CardTitle>
+        <CardDescription>{tAuth("continue-gpt")}</CardDescription>
+      </CardHeader>
+      <CardContent>
         {checkGithub() && (
           <Button
-            className="select-none"
-            size="lg"
-            icon={
+            variant="outline"
+            className="w-full justify-start gap-4 mb-2"
+            disabled={loadingGithub}
+            onClick={onGithubLogin}
+          >
+            {loadingGithub ? (
+              <Loader2 className="h-4 w-4 animate-spin" />
+            ) : (
               <Icon
                 className="text-black dark:text-white"
                 icon="github_line"
-                size={16}
+                size={20}
               />
-            }
-            loading={loadingGithub}
-            onClick={onGithubLogin}
-          >
+            )}
             {tAuth("continue-with-github")}
           </Button>
         )}
         {checkGoogle() && (
           <Button
-            className="select-none"
-            size="lg"
-            icon={<Icon icon="google_line" size={16} />}
-            loading={loadingGoogle}
+            variant="outline"
+            className="w-full justify-start gap-4"
+            disabled={loadingGoogle}
             onClick={onGoogleLogin}
           >
+            {loadingGoogle ? (
+              <Loader2 className="h-4 w-4 animate-spin" />
+            ) : (
+              <Icon icon="google_line" size={20} />
+            )}
             {tAuth("continue-with-google")}
           </Button>
         )}
-      </div>
-      {checkEmail() && (checkGithub() || checkGoogle()) && (
-        <Divider className="my-8">
-          <span className="text-sm uppercase">{tGlobal("or")}</span>
-        </Divider>
-      )}
-      {checkEmail() && (
-        <div className="flex flex-col gap-4">
-          <div>
-            <div className="font-medium mb-1 text-[13px]">
-              {tAuth("email-address")}
-            </div>
-            <Input
-              ref={inputRef}
-              size="lg"
-              placeholder={tGlobal("please-enter")}
-              allowClear
-              value={email}
-              onChange={setEmail}
-              onEnter={onLogin}
-            />
+        {checkEmail() && (checkGithub() || checkGoogle()) && (
+          <div className="flex items-center my-8">
+            <div className="h-[1px] flex-1 bg-[rgba(0,0,0,0.16)]"></div>
+            <div className="mx-4 text-[13px]">{tGlobal("or")}</div>
+            <div className="h-[1px] flex-1 bg-[rgba(0,0,0,0.16)]"></div>
           </div>
-          <Button
-            type="primary"
-            size="lg"
-            loading={loadingEmial}
-            onClick={onLogin}
-          >
-            {tGlobal("continue")}
-          </Button>
-        </div>
-      )}
-    </div>
+        )}
+        {checkEmail() && (
+          <>
+            <div className="flex flex-col space-y-1.5 mb-4">
+              <Label htmlFor="name">{tAuth("email-address")}</Label>
+              <Input
+                ref={inputRef}
+                id="name"
+                className="h-[38px]"
+                value={email}
+                onChange={(e) => setEmail(e.target.value)}
+              />
+            </div>
+            <Button
+              className="w-full"
+              onClick={onLogin}
+              disabled={loadingEmail}
+            >
+              {loadingEmail && (
+                <Loader2 className="mr-2 h-4 w-4 animate-spin" />
+              )}
+              {tGlobal("continue")}
+            </Button>
+          </>
+        )}
+      </CardContent>
+    </Card>
   );
 }